Chore: 타입 분리 및 수정
