+++ src/app/types/auth/authTypes.ts
@@ -1,0 +1,5 @@
+//로그인
+export interface PostAuth {
+  email: string;
+  password: string;
+}
